+++ ChangeLog.md
@@ -1,5 +1,15 @@
 # ChangeLog for Related
 
+## Unreleased
+
+
+
+## 1.6
+- FIX : PHP 8 - *19/08/2022* - 1.6.2
+- FIX : Compatibility V16 - newToken and family - *06/06/2022* - 1.6.1
+- NEW : Ajout d'une permission de lier / délier des éléments particuliers *23/05/2022* - 1.6.0
+- NEW : Ajout de la class TechATM pour l'affichage de la page "A propos" *11/05/2022* 1.5.0
+
 ## 1.4
 - FIX : Related bloc on projects cards is always empty *06/10/2022* - 1.4.10
 - FIX : Get unknown class from where it is already instanciated *18/05/2022* - 1.4.9